refactor: 리스트, 카드 구조 개선
