+++ src/components/features/list/MeetingList.tsx
@@ -9,7 +9,7 @@ import {
   useJoinedMeetingList,
 } from "@/hooks/queries/useMeetingInfiniteList";
 import { useUrlQuery } from "@/hooks/useUrlQuery";
-import { SortOrder, MeetingSortBy } from "@/types";
+import { SortOrder, MeetingSortBy, JoinedMeeting } from "@/types";
 import { NoResultFound } from "@/components/ui/NoResultFound";
 
 interface MeetingsClientProps {
@@ -41,7 +41,6 @@ export default function MeetingList({
 
   const {
     meetingList,
-    isLoading,
     fetchNextPage,
     hasNextPage,
     isFetchingNextPage,
@@ -56,18 +55,28 @@ export default function MeetingList({
     isFetchingNextPage,
   );
 
+  function isMeetingClosed(item: JoinedMeeting) {
+    return item.participantCount >= item.capacity;
+  }
+
+  const visibleMeetingList =
+    sortValue === "registrationEnd"
+      ? meetingList.filter((item) => !isMeetingClosed(item))
+      : meetingList;
+
   return (
     <div className="w-full">
       <div className="flex flex-col gap-6 sm:gap-8 lg:grid lg:grid-cols-2 lg:gap-x-10 lg:gap-y-12">
-        {meetingList.length > 0 ? (
-          <MeetingCard
-            meetingList={meetingList}
-            isLoading={isLoading}
-            onItemClick={(item) => router.push(`/meetings/${item.id}`)}
-            sortValue={sortValue}
-            onHeartClick={(item) => toggleFavorite(item)}
-            meetingStatusBadgeVisible={meetingStatusBadgeVisible}
-          />
+        {visibleMeetingList.length > 0 ? (
+          visibleMeetingList.map((item) => (
+            <MeetingCard
+              key={item.id}
+              item={item}
+              onItemClick={() => router.push(`/meetings/${item.id}`)}
+              onHeartClick={() => toggleFavorite(item)}
+              meetingStatusBadgeVisible={meetingStatusBadgeVisible}
+            />
+          ))
         ) : (
           <NoResultFound />
         )}
