+++ src/components/features/list/PostList.tsx
@@ -8,7 +8,6 @@ import { useRouter } from "next/navigation";
 import { Loader2 } from "lucide-react";
 import { useIntersectionObserver } from "@/hooks/useIntersectionObserver";
 import { cn } from "@/lib/utils";
-import PostCardListSkeleton from "@/components/skeleton/PostCardListSkeleton";
 import { QUERY_KEYS } from "@/constans/queryKey";
 import { getNextPageParam } from "@/lib/pagination";
 import { useUrlQuery } from "@/hooks/useUrlQuery";
@@ -48,8 +47,6 @@ export default function PostList() {
     isFetchingNextPage,
   );
 
-  if (isLoading) return <PostCardListSkeleton />;
-
   return (
     <div
       className={cn(
@@ -77,7 +74,7 @@ export default function PostList() {
                   "animate-fade-up group cursor-pointer overflow-hidden rounded-[24px] bg-white transition-all duration-300",
                   "border border-slate-100/50 shadow-[0_8px_30px_rgb(0,0,0,0.04)]",
                   "sm:hover:-translate-y-1 sm:hover:shadow-[0_20px_40px_rgba(38,6,86,0.08)]",
-                  "focus-visible:ring-black focus-visible:ring-2 focus-visible:outline-none",
+                  "focus-visible:ring-2 focus-visible:ring-black focus-visible:outline-none",
                 )}
               >
                 <PostCard
